🍒 [Dashboard filters] Create initial infrastructure and group repros (#17082)
* [Dashboard filters] Create initial infrastructure and group repros (#17079)

* Extract repro for #15279 into a separate file

* Extract repro for #15689 into a separate file

* Extract repro for #15695 into a separate file

* Extract repro for #15119 into a separate file

* Extract repro for #12614 into a separate file

* Extract repro for #12720 into a separate file

* Extract repro for #6660 into a separate file

* Extract repro for #16103 into a separate file

* Extract repro for #13150 into a separate file

* Move existing tests related to the dashboard filters into a new folder

* Create new CI group `dashboard-filters`

* Skip test that relies on a refactor from `master`
